feat: Implemented Create Todo function
Create Todo function has been implemented completely. Tests will be run and bugs will be fixed once all the predefined functions have been implemented.
I have added two new folders named 'businessLogic' and 'dataLayer'. Inside of 'businessLogic' I have created a file named 'todos.ts'. This file contains the actual business logic of how to create a new item, whereas the handler function inside 'createTodo.ts' is only used to receive the request. This is done so that the business logic can be ported easily to different FaaS providers, therefore minimizing the risk of a hard vendor-lockin.
Additionally, the business logic and the data layer have been separated, so that the business logic only focusses on how to do its job, not caring with what kind of data storage technology is used in the background. Therefore, files inside the 'dataLayer' folder implement the functionality on how data is processed and stored in the chosen database. In our case, one file ('todosAccess.ts') represents one DynamoDB table on which we want to run operations against. If we want to change the database technology in the future, all we have to do is to change the implementation of the dataLayer file. The business logic won't (at least should) not be affected by these changes.