fix: Fixed DeleteTodo function
Missing 'await' keyword broke the DeleteTodo function, as we have expected a boolean value when checking for if the user is allowed to access a certain todo. Because this is done asynchronous, the function returns a promise expecting a boolean value. Without the 'await' keyword we therefore have always gotten a Promise object, which inside an if statement is evaluated as a truthy value. Therefore, all requests have returned a status code of 204 (which means the deletion was successful) instead of a 403 forbidden (which means that the user has insufficient permissions to access the specified todo)